requirements.txt
accelerate==1.8.1
av==14.3.0
colorama==0.4.6
diffusers==0.32.2
easydict==1.13
einops==0.8.1
huggingface_hub==0.30.2
hydra-core==1.3.2
imageio==2.37.0
internetarchive
lightning==2.5.1
lightning_utilities==0.14.3
matplotlib==3.10.3
numpy==2.2.6
omegaconf==2.3.0
opencv_python==4.11.0.86
pandas==2.3.1
Pillow==11.3.0
pyiqa==0.1.10
pytubefix==8.12.3
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
roma==1.5.2.1
rotary_embedding_torch==0.8.6
safetensors==0.5.3
scikit_learn
scipy==1.15.3
seaborn==0.13.2
scikit-image
timm==1.0.17
torch==2.6.0
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.1
typing_extensions==4.14.1
wandb_osh==1.2.2
yt_dlp==2025.6.30
